requirements.txt
fastapi
uvicorn
httpx
pydantic
google-generativeai
python-dotenv